moving childern of DraggableItem to PopUp.js
