+++ client/src/components/DraggableItem.js
@@ -1,8 +1,6 @@
 import _ from 'lodash';
 import React, { Component } from 'react';
 import { DragSource, DropTarget } from 'react-dnd';
-import IconBar from 'react-icons/lib/fa/bars';
-import styled from 'styled-components';
 
 const ItemTypes = {
   ROLE: 'role'
@@ -47,8 +45,6 @@ export const DraggableItem = _.flow([
   class DraggableItem extends Component {
     render() {
       const {
-        name,
-        onChange,
         connectDragSource,
         connectDropTarget,
         isDragging,
@@ -56,7 +52,7 @@ export const DraggableItem = _.flow([
         canDrop
       } = this.props;
 
-      let isDragEntering;
+      let isDragEntering = false;
       if (isOver && canDrop) {
         isDragEntering = true;
       }
@@ -64,74 +60,15 @@ export const DraggableItem = _.flow([
       return connectDropTarget(
         connectDragSource(
           <div>
-            <Wrapper>
-              <StyleIconBar />
-              <StyleInput
-                data-hj-whitelist
-                type="text"
-                value={name}
-                onChange={onChange}
-                isDragging={isDragging}
-                isDragEntering={isDragEntering}
-              />
-              {this.props.children}
-            </Wrapper>
+            {React.cloneElement(this.props.children, {
+              isDragging: isDragging,
+              isDragEntering: isDragEntering
+            })}
           </div>
         )
       );
     }
   }
 );
-const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-Wrapper.displayName = 'Wrapper';
-
-const StyleIconBar = styled(IconBar)`
-  cursor: move;
-  margin-right: 5px;
-`;
-StyleIconBar.displayName = 'StyleIconBar';
-
-const StyleInput = styled.input`
-  opacity: ${props => (props.isDragging ? 0.5 : 1)};
-  background-color: ${props => (props.isDragEntering ? '#c1c1c1' : '#fff')};
-  border: solid 1px #c1c1c1;
-  border-radius: 4px;
-  box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.1);
-  box-sizing: border-box;
-  color: #333;
-  padding: 9px;
-  &.extra-space-for-addon {
-    padding-left: 48px;
-  }
-  &:hover {
-    border-color: #8b8b8b;
-  }
-  &:focus {
-    border-color: #64b8dc;
-    box-shadow: none;
-    outline: 0;
-  }
-  &[disabled] {
-    background: #e5e5e5;
-    border-color: #c1c1c1;
-    box-shadow: none;
-  }
-  &::-webkit-input-placeholder {
-    color: #999;
-  }
-  &:-moz-placeholder {
-    color: #999;
-  }
-  &::-moz-placeholder {
-    color: #999;
-  }
-  &:-ms-input-placeholder {
-    color: #999;
-  }
-`;
-StyleInput.displayName = 'StyleInput';
 
 export default DraggableItem;
